requirements.txt
langgraph
langchain
langchain-openai
langchain-community
google-search-results
langchain-core
rdkit
requests
openai
python-dotenv